+++ Editor/Watchers/BoolVariableWatcherEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Watchers.Editor
+{
+    [CustomEditor(typeof(BoolVariableWatcher), true)]
+    public class BoolVariableWatcherEditor : RuntimeVariableWatcherEditor<bool, BoolVariable> { }
+}