+++ Editor/Watchers/StringVariableWatcherEditor.cs
@@ -1,0 +1,7 @@
+using UnityEditor;
+
+namespace UnderLogic.Variables.Watchers.Editor
+{
+    [CustomEditor(typeof(StringVariableWatcher), true)]
+    public class StringVariableWatcherEditor : RuntimeVariableWatcherEditor<string, StringVariable> { }
+}